refactor(domain): User 도메인 구조 변경 및 도메인의 import 추가
