+++ src/main/java/com/example/demo/domain/chat/Chat.java
@@ -1,8 +1,16 @@
 package com.example.demo.domain.chat;
 
-import com.example.demo.domain.User;
 import com.example.demo.domain.common.BaseEntity;
-import jakarta.persistence.*;
+import com.example.demo.domain.user.User;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 
 @Entity
 @Table(name = "chat")
