+++ src/main/java/com/example/demo/domain/mission/CustomMission.java
@@ -1,6 +1,6 @@
 package com.example.demo.domain.mission;
 
-import com.example.demo.domain.User;
+import com.example.demo.domain.user.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
