+++ src/main/java/com/example/demo/domain/user/User.java
@@ -1,7 +1,17 @@
-package com.example.demo.domain;
+package com.example.demo.domain.user;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
+@Getter
+@NoArgsConstructor
 @Entity
 @Table(name = "Users")
 public class User {
@@ -22,6 +32,16 @@ public class User {
     @Column(nullable = false)
     private String refreshToken;
 
-    protected User() {
+    @Builder
+    public User(String name, String email, Integer point, String refreshToken) {
+        this.name = name;
+        this.email = email;
+        this.point = point;
+        this.refreshToken = refreshToken;
+    }
+
+    public User update(String name) {
+        this.name = name;
+        return this;
     }
 }
