+++ src/main/java/com/example/demo/domain/userEmotion/UserEmotionHistory.java
@@ -1,7 +1,7 @@
 package com.example.demo.domain.userEmotion;
 
-import com.example.demo.domain.User;
 import com.example.demo.domain.common.BaseEntity;
+import com.example.demo.domain.user.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
